general code cleanup
