+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/s3/S3ExplorerNode.kt
@@ -16,6 +16,7 @@ import software.aws.toolkits.jetbrains.core.explorer.AwsExplorerService
 import software.aws.toolkits.jetbrains.core.explorer.nodes.AwsExplorerNode
 import software.aws.toolkits.jetbrains.core.explorer.nodes.AwsExplorerResourceNode
 import software.aws.toolkits.jetbrains.core.explorer.nodes.AwsExplorerServiceRootNode
+import software.aws.toolkits.jetbrains.services.s3.bucketEditor.S3VirtualBucket
 import software.aws.toolkits.jetbrains.services.s3.resources.S3Resources
 import software.aws.toolkits.jetbrains.services.telemetry.TelemetryService
 
@@ -43,7 +44,9 @@ class S3BucketNode(project: Project, val bucket: Bucket) :
             val editorManager = FileEditorManager.getInstance(nodeProject)
             // See if there is already an open editor, otherwise make a new one
             val virtualFile =
-                editorManager.openFiles.firstOrNull { (it as? S3VirtualBucket)?.s3Bucket?.equals(bucket) == true } ?: S3VirtualBucket(bucket)
+                editorManager.openFiles.firstOrNull { (it as? S3VirtualBucket)?.s3Bucket?.equals(bucket) == true } ?: S3VirtualBucket(
+                    bucket
+                )
             editorManager.openTextEditor(OpenFileDescriptor(nodeProject, virtualFile), true)
             recordOpenTelemetry()
         }
