+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/s3/S3ViewerProvider.kt
@@ -17,6 +17,7 @@ import com.intellij.openapi.vfs.VirtualFile
 import software.amazon.awssdk.services.s3.S3Client
 import software.aws.toolkits.jetbrains.core.AwsClientManager
 import software.aws.toolkits.jetbrains.services.s3.bucketEditor.S3ViewerPanel
+import software.aws.toolkits.jetbrains.services.s3.bucketEditor.S3VirtualBucket
 import java.beans.PropertyChangeListener
 import javax.swing.JComponent
