+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/s3/bucketActions/CreateBucketAction.kt
@@ -1,6 +1,6 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0package software.aws.toolkits.jetbrains.services.s3.BucketActions
-package software.aws.toolkits.jetbrains.services.s3
+package software.aws.toolkits.jetbrains.services.s3.bucketActions
 
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.LangDataKeys
@@ -8,6 +8,7 @@ import icons.AwsIcons
 import software.amazon.awssdk.services.s3.S3Client
 import software.aws.toolkits.jetbrains.components.telemetry.AnActionWrapper
 import software.aws.toolkits.jetbrains.core.AwsClientManager
+import software.aws.toolkits.jetbrains.services.s3.CreateS3BucketDialog
 import software.aws.toolkits.resources.message
 
 class CreateBucketAction : AnActionWrapper(message("s3.create.bucket.title"), null, AwsIcons.Resources.S3_BUCKET) {
