+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/s3/bucketEditor/S3ColumnInfo.kt
@@ -5,7 +5,7 @@ package software.aws.toolkits.jetbrains.services.s3.bucketEditor
 import com.intellij.openapi.util.text.StringUtil
 import com.intellij.ui.treeStructure.treetable.TreeTableModel
 import com.intellij.util.ui.ColumnInfo
-import software.aws.toolkits.jetbrains.services.s3.S3Resources
+import software.aws.toolkits.jetbrains.services.s3.resources.S3Resources
 import software.aws.toolkits.resources.message
 import javax.swing.tree.DefaultMutableTreeNode
