+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/s3/bucketEditor/S3VirtualBucket.kt
@@ -1,7 +1,7 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.services.s3
+package software.aws.toolkits.jetbrains.services.s3.bucketEditor
 
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.testFramework.LightVirtualFile
