+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/s3/objectActions/CopyPathAction.kt
@@ -7,14 +7,14 @@ import com.intellij.icons.AllIcons
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.ide.CopyPasteManager
 import software.aws.toolkits.jetbrains.components.telemetry.ActionButtonWrapper
-import software.aws.toolkits.jetbrains.services.s3.S3VirtualBucket
 import software.aws.toolkits.jetbrains.services.s3.bucketEditor.S3TreeContinuationNode
 import software.aws.toolkits.jetbrains.services.s3.bucketEditor.S3TreeTable
+import software.aws.toolkits.jetbrains.services.s3.bucketEditor.S3VirtualBucket
 import software.aws.toolkits.resources.message
 import java.awt.datatransfer.StringSelection
 
 class CopyPathAction(
-    private var treeTable: S3TreeTable,
+    private val treeTable: S3TreeTable,
     val bucket: S3VirtualBucket
 ) : ActionButtonWrapper(message("s3.copy.path"), null, AllIcons.Actions.Copy) {
     override fun isEnabled(): Boolean = treeTable.selectedRows.size == 1 && !treeTable.getSelectedNodes().any { it is S3TreeContinuationNode }
