+++ jetbrains-core/tst/software/aws/toolkits/jetbrains/services/s3/RenameObjectTest.kt
@@ -18,6 +18,7 @@ import software.amazon.awssdk.services.s3.model.DeleteObjectResponse
 import software.aws.toolkits.jetbrains.core.MockClientManagerRule
 import software.aws.toolkits.jetbrains.services.s3.bucketEditor.S3TreeObjectNode
 import software.aws.toolkits.jetbrains.services.s3.bucketEditor.S3TreeTable
+import software.aws.toolkits.jetbrains.services.s3.bucketEditor.S3VirtualBucket
 import software.aws.toolkits.jetbrains.services.s3.objectActions.RenameObjectAction
 import software.aws.toolkits.jetbrains.utils.delegateMock
 import java.time.Instant
